create copyCat()
Finally figured out a way to transfer the content of #ae into #be, as the user types, but each letter is followed by an array comma. Can't get the .join() method to work, I guess because, as MDN says, "If the array has only one item, then that item will be returned without using the separator." And technically, each time the mimic array is displayed, it has only one item. Still trying to figure out a solution.