update copyCat() solution
Okay, yayayay omg it's been, what, a week?? More?

After a bunch of trial and error, I've  f i n a l l y
 figured out a way to transfer the text from #ae to #be as it's typed. I'm going to commit a few of the commented-out failures, so I can look back on them later.

Some notes! I considered string methods like .substring(), when I realized the single-item array was preventing me from using .join() to remove the array commas separating each letter, but strings are immutable, so that was a dud.

So, I thought I'd try to add a random letter into the mimic array, with a for loop, so that could be joined with the spliced letters array letter. Then I tried accessing the desired letter with bracket notation, instead of splicing the "whole" one-item array, and I guess doing so finally enabled me to use the .join() method. Should have figured that out sooner.