update #ae .split() with regex
Updated the `.split()` method, so the string now splits at non-word characters. Using `\W` splits at spaces, and most important, also punctuation that impeded matches between typed words and the phrases in the translations array. Note: the capturing parentheses in the regex include the split characters in the #beSplit array, so they appear in the British translation.