fix translated prevWord bug
Before the translations included additional divs and spans, I was able to use a regex to strip the `<mark>` tags to get the translated word.

I tried to figure out a new regex that would remove all tags and innerHTML except the desired word, but that was a bust. I also read up on DOMParser, but in the end it just seemed easiest to turn the string into a jQuery object, find the mark, extract the text, and update prevWord accordingly.