create oneOfTwoCapitalized
The condition that was supposed to match `capitalized` two-word array phrases (e.g., "Two thirty" to "Half two") was accidentally capitalizing single array words that followed non-array words (e.g., "The cat" became "The Moggy").

To fix this, I created the variable `oneOfTwoCapitalized`, which should be true, for array phrases like "Two thirty," while `capitalized` is now false. Then I created a new condition in the caseMatch() function that checks whether `oneOfTwoCapitalized` is true, and capitalizes the first letter of the first word in such phrases.